[SPARK-44005][PYTHON] Improve error messages for regular Python UDTFs that return non-tuple values
### What changes were proposed in this pull request?

This PR improves error messages for regular Python UDTFs when the result rows are not one of tuple, list and dict. Note this is supported when arrow optimization is enabled.

### Why are the changes needed?

To make Python UDTFs more user friendly.

### Does this PR introduce _any_ user-facing change?

Yes.
```
class TestUDTF:
    def eval(self, a: int):
        yield a
```
Before this PR, this will fail with this error
`Unexpected tuple 1 with StructType`

After this PR, this will have a more user-friendly error:
`[UDTF_INVALID_OUTPUT_ROW_TYPE] The type of an individual output row in the UDTF is invalid. Each row should be a tuple, list, or dict, but got 'int'. Please make sure that the output rows are of the correct type.`

### How was this patch tested?

Existing UTs.

Closes #42353 from allisonwang-db/spark-44005-non-tuple-return-val.

Authored-by: allisonwang-db <allison.wang@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>