[SPARK-44741][CORE] Support regex-based MetricFilter in `StatsdSink`
### What changes were proposed in this pull request?

Adding additional option in StatsdSink to allow passing regex as a filter option for reporting metrics
Added unit test case to validate the application of filter in the StatsdSink
Updated metrics config to represent the added config for StatsdSink

### Why are the changes needed?

In the current state Spark metrics instances send a large range of metrics that are useful for debugging and performance analysis. There are cases where the consumers would like to switch between deeper metrics and only specific custom metrics that are required for system maintenance. This option introduced in StatsdSink extends the existing filtering option in the reporter and exposes the same as an option. This implementation is similar to GraphiteSink that supports a similar feature.

### Does this PR introduce _any_ user-facing change?

Yes , This PR exposes an additional config that users can provide optionally to filter the metrics that are being reported. This is a optional field and defaulted to the existing behaviour where the sink reports all the metrics.

### How was this patch tested?

Unit tests has been added in the PR to validate the change

Closes #42416 from ramesh-muthusamy/SPARK-44741.

Authored-by: Rameshkrishnan Muthusamy <rameshkrishnan_muthusamy@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>