[SPARK-44737][SQL][UI] Should not display json format errors on SQL page for non-SparkThrowables on SQL Tab
### What changes were proposed in this pull request?

This pull request addresses the issue of inconsistent display for SparkThrowables and others. Currently, SparkThrowables appear as plain text while others are displayed in JSON format on SQL Tab. This fix aims to make the display consistent across all items.

### Why are the changes needed?

To fix the inconsistent behavior, and the JSON format is not correctly displayed.

For those non-SparkThrowables thrown by SQLExecution may contain errors either that we are not ready covered by the new error handling framework or some critical path(what I have met is OOM in RDD.getPartitions), we shall supply the stacktraces

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

add a unit test

Closes #42407 from yaooqinn/SPARK-44737.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>