[SPARK-44795][CONNECT] CodeGenerator Cache should be classloader specific
### What changes were proposed in this pull request?
When you currently use a REPL generated class in a UDF you can get an error saying that that class is not equal to that class. This error is thrown in a code generated class. The problem is that the classes have been loaded by different classloaders. We cache generated code and use the textual code as the string. The problem with this is that in Spark Connect users are free in supplying user classes that can have arbitrary names, a name can point to an entirely different class, or it can point to the same class (in bytes) but loaded by a different classloader. We need to add the classloader to the key in the code gen cache to make this safe.

There are roughly two ways how this problem can arise:
1. Two sessions use the same class names. This is particularly easy when you use the REPL because this  always generates the same names.
2. You run in single process mode. In this case wholestage codegen will test compile the class using a different classloader then the 'executor', while sharing the same code generator cache.

### Why are the changes needed?
We want to be able to use REPL (and other)

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
I added a test to the `ReplE2ESuite`.

Closes #42478 from hvanhovell/SPARK-44795.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>