[SPARK-43885][SQL][FOLLOWUP] Instruction#dataType should not fail
### What changes were proposed in this pull request?

This is a followup of https://github.com/apache/spark/pull/41448 . As an optimizer rule, the produced plan should be resolved and resolved expressions should be able to report data type. The `Instruction` expression fails to report data type and may break external optimizer rules. This PR makes it to return dummy NullType.

### Why are the changes needed?

to not break external optimizer rules.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

existing tests

Closes #42482 from cloud-fan/merge.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>