[SPARK-44904][PYTHON][DOCS] Correct the `versionadded` of `sql.functions.approx_percentile` to 3.5.0
### What changes were proposed in this pull request?
`sql.functions.approx_percentile`  was introduced in SPARK-43941 (https://github.com/apache/spark/pull/41588). This is a pr for the Spark 3.5.0 and it does not belong to Spark 3.4.0. Therefore, this pr corrects the `versionadded` of `sql.functions.approx_percentile` to 3.5.0 and removed `versionchanged`.

### Why are the changes needed?
Correct the `versionadded` of `sql.functions.approx_percentile`.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass GitHub Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #42597 from LuciferYang/SPARK-44904.

Lead-authored-by: yangjie01 <yangjie01@baidu.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>